checkpoint: all changes except compound fold; tests outstanding
