+++ tests/test_core_catalog.py
